+++ libraries/MTP40F/MTP40F.h
@@ -3,7 +3,7 @@
 //    FILE: MTP40F.h
 //  AUTHOR: Rob Tillaart
 //    DATE: 2023-07-25
-// VERSION: 0.1.0
+// VERSION: 0.1.1
 // PURPOSE: Arduino library for MTP40F CO2 sensor
 //     URL: https://github.com/RobTillaart/MTP40F
 
@@ -11,7 +11,7 @@
 #include "Arduino.h"
 
 
-#define MTP40F_LIB_VERSION                    (F("0.1.0"))
+#define MTP40F_LIB_VERSION                    (F("0.1.1"))
 
 
 #define MTP40F_DEFAULT_ADDRESS                0x64
@@ -20,17 +20,18 @@
 #define MTP40F_INVALID_AIR_PRESSURE           0x01
 #define MTP40F_INVALID_GAS_LEVEL              0x02
 #define MTP40F_INVALID_ADDRESS                0xFF
+#define MTP40F_REQUEST_FAILED                 0xFFFF
 
 
 class MTP40F
 {
 public:
-  MTP40F(Stream * str);
+  MTP40F(Stream * stream);
 
   bool     begin();
 
   int      getAirPressureReference();
-  bool     setAirPressureReference(int apr);
+  bool     setAirPressureReference(int apr = 1013);
 
   uint32_t getGasConcentration();  //  returns PPM
 
@@ -45,18 +46,18 @@ public:
   bool     openSelfCalibration();
   bool     closeSelfCalibration();
   uint8_t  getSelfCalibrationStatus();
-  bool     setSelfCalibrationHours(uint16_t hrs);
+  bool     setSelfCalibrationHours(uint16_t hours = 168);
   uint16_t getSelfCalibrationHours();
 
 
   //  set timeout of serial communication.
-  void     setTimeout(uint32_t to = 100) { _timeout = to; };
+  void     setTimeout(uint32_t timeout = 100) { _timeout = timeout; };
   uint32_t getTimeout() { return _timeout; };
 
   uint32_t lastRead() { return _lastRead; };
 
-  //       2 = MTP40C
-  //       3 = MTP40D
+  //       2 = MTP40C - other library
+  //       3 = MTP40D - other library
   //       5 = MTP40F
   uint8_t  getType()  { return _type; };
   int      lastError();
@@ -79,8 +80,8 @@ protected:
   bool     _suppressError = false;
   int      _lastError     = MTP40F_OK;
 
-  bool     request(uint8_t *data, uint8_t cmdlen, uint8_t anslen);
-  uint16_t CRC(uint8_t *data, uint16_t len);
+  bool     request(uint8_t *data, uint8_t commandLength, uint8_t responseLength);
+  uint16_t CRC(uint8_t *data, uint16_t length);
 };