+++ libraries/MTP40F/CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.1] - 2023-08-02
+- fix **getGasConcentration()** status handling
+- update readme.md
+- improve readability some parameters
+- add (SoftwareSerial) schema to several examples
+- add **MTP40F_REQUEST_FAILED** (error handling is still to do)
+- add default value to **setAirPressureReference(apr = 1013)**
+- add default value to **setSelfCalibrationHours(hours = 168)**
+- update examples
+
+
 ## [0.1.0] - 2023-07-25
 - initial version.